Created Sort as seperate IT class
